reconstruct handlers and services
