+++ web/index.html
@@ -62,7 +62,6 @@
                 </div>
                 <div class="flex items-center space-x-4">
                     <div class="flex items-center space-x-2">
-                        <span class="text-sm" data-i18n="nav.status">状态:</span>
                         <span id="server-status" class="px-2 py-1 rounded text-xs font-medium bg-gray-500" data-i18n="nav.status.unknown">未知</span>
                     </div>
                     <div class="language-toggle">
